[22/21][eas-cli] Move some files around to avoid private functions
