+++ packages/eas-cli/src/commandUtils/context/DynamicProjectConfigContextField.ts
@@ -2,8 +2,8 @@ import { ExpoConfig } from '@expo/config-types';
 
 import { ExpoConfigOptions, getExpoConfig } from '../../project/expoConfig';
 import ContextField, { ContextOptions } from './ContextField';
-import ProjectConfigContextField from './ProjectConfigContextField';
-import ProjectDirContextField from './ProjectDirContextField';
+import { findProjectDirAndVerifyProjectSetupAsync } from './contextUtils/findProjectDirAndVerifyProjectSetupAsync';
+import { getProjectIdAsync } from './contextUtils/getProjectIdAsync';
 
 export type DynamicConfigContextFn = (options?: ExpoConfigOptions) => Promise<{
   projectId: string;
@@ -14,9 +14,9 @@ export type DynamicConfigContextFn = (options?: ExpoConfigOptions) => Promise<{
 export class DynamicProjectConfigContextField extends ContextField<DynamicConfigContextFn> {
   async getValueAsync({ nonInteractive }: ContextOptions): Promise<DynamicConfigContextFn> {
     return async (options?: ExpoConfigOptions) => {
-      const projectDir = await ProjectDirContextField['findProjectDirAndVerifyProjectSetupAsync']();
+      const projectDir = await findProjectDirAndVerifyProjectSetupAsync();
       const expBefore = getExpoConfig(projectDir, options);
-      const projectId = await ProjectConfigContextField['getProjectIdAsync'](expBefore, {
+      const projectId = await getProjectIdAsync(expBefore, {
         nonInteractive,
         env: options?.env,
       });
